Consolidated setupOpenMRS functions into one.
