--- src/components/choose-locale/change-locale-offline.resource.tsx
@@ -1,6 +1,0 @@
-import { PrimaryNavigationDb } from '../../offline';
-
-export function postUserPropertiesOffline(userUuid: string, userProperties: any): Promise<any> {
-  const db = new PrimaryNavigationDb();
-  return db.userPropertiesChanges.add({ userUuid, changes: userProperties });
-}
