--- src/components/user-panel-switcher-item/user-panel-switcher-offline.component.tsx
@@ -1,8 +1,0 @@
-import React from 'react';
-import UserPanelSwitcher, { UserPanelSwitcherItemProps } from './user-panel-switcher.component';
-
-const UserPanelSwitcherOffline: React.FC<UserPanelSwitcherItemProps> = props => (
-  <UserPanelSwitcher {...props} isLogoutEnabled={false} />
-);
-
-export default UserPanelSwitcherOffline;
