--- src/components/user-panel-switcher-item/user-panel-switcher-online.component.tsx
@@ -1,8 +1,0 @@
-import React from 'react';
-import UserPanelSwitcher, { UserPanelSwitcherItemProps } from './user-panel-switcher.component';
-
-const UserPanelSwitcherOnline: React.FC<UserPanelSwitcherItemProps> = props => (
-  <UserPanelSwitcher {...props} isLogoutEnabled={true} />
-);
-
-export default UserPanelSwitcherOnline;
