+++ src/index.ts
@@ -1,4 +1,5 @@
 import { defineConfigSchema, getAsyncLifecycle } from '@openmrs/esm-framework';
+import { postUserPropertiesOnline, postUserPropertiesOffline } from './components/choose-locale/change-locale.resource';
 import { configSchema } from './config-schema';
 
 const backendDependencies = { 'webservices.rest': '^2.2.0' };
@@ -15,48 +16,42 @@ function setupOpenMRS() {
   defineConfigSchema(moduleName, configSchema);
 
   return {
-    lifecycle: getAsyncLifecycle(() => import('./root-online.component'), options),
-    activate: (location: Location) => !location.pathname.startsWith(window.getOpenmrsSpaBase() + 'login'),
+    pages: [
+      {
+        load: getAsyncLifecycle(() => import('./root.component'), options),
+        route: (location: Location) => !location.pathname.startsWith(window.getOpenmrsSpaBase() + 'login'),
+        online: { syncUserPropertiesChangesOnLoad: true },
+        offline: { syncUserPropertiesChangesOnLoad: false },
+      },
+    ],
     extensions: [
       {
         id: 'default-user-panel',
         slot: 'user-panel-slot',
         load: getAsyncLifecycle(
-          () => import('./components/user-panel-switcher-item/user-panel-switcher-online.component'),
+          () => import('./components/user-panel-switcher-item/user-panel-switcher.component'),
           options,
         ),
+        online: {
+          isLogoutEnabled: true,
+        },
+        offline: {
+          isLogoutEnabled: false,
+        },
       },
       {
         id: 'change-locale',
         slot: 'user-panel-slot',
-        load: getAsyncLifecycle(() => import('./components/choose-locale/change-locale-online.component'), options),
+        load: getAsyncLifecycle(() => import('./components/choose-locale/change-locale.component'), options),
+        online: {
+          postUserProperties: postUserPropertiesOnline,
+        },
+        offline: {
+          postUserProperties: postUserPropertiesOffline,
+        },
       },
     ],
   };
 }
 
-function setupOpenMRSOffline() {
-  defineConfigSchema(moduleName, configSchema);
-
-  return {
-    lifecycle: getAsyncLifecycle(() => import('./root-offline.component'), options),
-    activate: (location: Location) => !location.pathname.startsWith(window.getOpenmrsSpaBase() + 'login'),
-    extensions: [
-      {
-        id: 'default-user-panel',
-        slot: 'user-panel-slot',
-        load: getAsyncLifecycle(
-          () => import('./components/user-panel-switcher-item/user-panel-switcher-offline.component'),
-          options,
-        ),
-      },
-      {
-        id: 'change-locale',
-        slot: 'user-panel-slot',
-        load: getAsyncLifecycle(() => import('./components/choose-locale/change-locale-offline.component'), options),
-      },
-    ],
-  };
-}
-
-export { backendDependencies, importTranslation, setupOpenMRS, setupOpenMRSOffline };
+export { backendDependencies, importTranslation, setupOpenMRS };
