--- src/root-offline.component.tsx
@@ -1,6 +1,0 @@
-import React from 'react';
-import Root from './root.component';
-
-export default function RootOnline() {
-  return <Root syncUserPropertiesChangesOnLoad={false} />;
-}
